+++ src/org/usfirst/frc/team166/robot/OI.java
@@ -4,10 +4,10 @@ import edu.wpi.first.wpilibj.Joystick;
 import edu.wpi.first.wpilibj.buttons.JoystickButton;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 
-import org.usfirst.frc.team166.robot.commands.ClimberOn;
-import org.usfirst.frc.team166.robot.commands.ToggleGearManip;
+import org.usfirst.frc.team166.robot.commands.Climber.ClimberOn;
 import org.usfirst.frc.team166.robot.commands.GearManipulator.CloseManipulator;
 import org.usfirst.frc.team166.robot.commands.GearManipulator.OpenManipulator;
+import org.usfirst.frc.team166.robot.commands.GearManipulator.ToggleGearManip;
 
 /**
  * This class is the glue that binds the controls on the physical operator interface to the commands and command groups
