+++ src/org/usfirst/frc/team166/robot/commands/GearManipulator/ToggleGearManip.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc.team166.robot.commands;
+package org.usfirst.frc.team166.robot.commands.GearManipulator;
 
 import edu.wpi.first.wpilibj.command.Command;
 
@@ -7,20 +7,18 @@ import org.usfirst.frc.team166.robot.Robot;
 /**
  *
  */
-public class ClimberOn extends Command {
+public class ToggleGearManip extends Command {
 
-	public ClimberOn() {
+	public ToggleGearManip() {
 		// Use requires() here to declare subsystem dependencies
 		// eg. requires(chassis);
 	}
 
-	// Called just before this Command runs the first time
 	@Override
 	protected void initialize() {
-		Robot.climber.climberOn();
+		Robot.gearManipulator.togglesGearmanip();
 	}
 
-	// Called repeatedly when this Command is scheduled to run
 	@Override
 	protected void execute() {
 	}
@@ -28,7 +26,7 @@ public class ClimberOn extends Command {
 	// Make this return true when this Command no longer needs to run execute()
 	@Override
 	protected boolean isFinished() {
-		return false;
+		return true;
 	}
 
 	// Called once after isFinished returns true
@@ -40,6 +38,5 @@ public class ClimberOn extends Command {
 	// subsystems is scheduled to run
 	@Override
 	protected void interrupted() {
-		Robot.climber.climberOff();
 	}
 }
